remove bolt_db.go file
